+++ scipy/optimize/tests/test__linprog_clean_inputs.py
@@ -6,7 +6,7 @@ from __future__ import division, print_function, absolute_import
 import numpy as np
 from numpy.testing import assert_, assert_allclose
 from pytest import raises as assert_raises
-from scipy.optimize._linprog import _clean_inputs
+from scipy.optimize._linprog_util import _clean_inputs
 from copy import deepcopy
