+++ scipy/optimize/tests/test_linprog.py
@@ -1157,9 +1157,11 @@ class TestLinprogSimplexCommon(BaseTestLinprogSimplex):
             assert_(res.pop('status') in range(4))
             assert_(isinstance(res.pop('nit'), int))
             assert_(isinstance(complete, bool))
+            assert_(isinstance(message, str))
 
             if complete:
                 last_cb['x'] = res['x']
+                last_cb['fun'] = res['fun']
                 last_cb['slack'] = res['slack']
                 last_cb['con'] = res['con']
 
@@ -1169,6 +1171,7 @@ class TestLinprogSimplexCommon(BaseTestLinprogSimplex):
         res = linprog(c, A_ub=A_ub, b_ub=b_ub, callback=cb, method=self.method)
 
         _assert_success(res, desired_fun=-18.0, desired_x=[2, 6])
+        assert_allclose(last_cb['fun'], res['fun'])
         assert_allclose(last_cb['x'], res['x'])
         assert_allclose(last_cb['con'], res['con'])
         assert_allclose(last_cb['slack'], res['slack'])
